+++ internal/models/user_data.go
@@ -3,6 +3,6 @@ package models
 type UserData struct {
 	UserId    string  `csv:"user_id"`
 	Currency  string  `csv:"currency"`
-	Timestamp int     `csv:"timestamp"`
+	Timestamp int64   `csv:"timestamp"`
 	Delta     float64 `csv:"delta"`
 }
